opsjobitem decimal, customer contract add last updated at and by

+++ app/Models/OpsJobTask.php
@@ -26,7 +26,7 @@ class OpsJobTask extends Model
     ];
 
     protected $casts = [
-        'sequence' => 'integer',
+        'sequence' => 'float',
         'value'    => 'integer',
         'qty'      => 'integer',
         'latitude' => 'float',
